Revert "Fix few cast issues."
This reverts commit 899ca05d24e4cb808d42070296dda13d1808fd6b.

+++ hardware/MySensorsBase.cpp
@@ -2022,7 +2022,7 @@ void MySensorsBase::ParseLine(const std::string &sLine)
 			bHaveValue = true;
 			break;
 		case V_DIRECTION:
-			pChild->SetValue(vType, (int)std::lrint(atof(payload.c_str())));
+			pChild->SetValue(vType, std::lrint(atof(payload.c_str())));
 			bHaveValue = true;
 			break;
 		case V_LIGHT_LEVEL:
